+++ CarRental.PresentationLayer/Enums/PitchComponent.cs
