+++ CarRental.PresentationLayer/Enums/PowerComponent.cs
