+++ CarRental.PresentationLayer/Enums/TypeTire.cs
